# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Radhika Khattar AI Persona API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ChatMessageInput {
  /** User message text */
  message: string;
  /** Session identifier for conversation continuity */
  sessionId: string;
}

export interface TimeSlot {
  /** ISO 8601 datetime */
  start: string;
  /** ISO 8601 datetime */
  end: string;
  /** Human-readable label e.g. "Mon Jun 9, 2-3 PM IST" */
  label: string;
}

export interface ChatResponse {
  reply: string;
  sessionId: string;
  /** @nullable */
  bookingSlots?: TimeSlot[] | null;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: number;
  sessionId: string;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface AvailabilityResponse {
  slots: TimeSlot[];
}

export interface BookingInput {
  /** ISO 8601 start datetime */
  start: string;
  /** ISO 8601 end datetime */
  end: string;
  attendeeEmail: string;
  attendeeName: string;
  /** @nullable */
  summary?: string | null;
}

export interface BookingConfirmation {
  eventId: string;
  /** @nullable */
  meetLink: string | null;
  start: string;
  end: string;
  message?: string;
}

export interface TwilioGatherInput {
  SpeechResult?: string;
  CallSid?: string;
  From?: string;
}

export type GetChatHistoryParams = {
sessionId: string;
};