# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Radhika Khattar AI Persona API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: chat
    description: Chat with Radhika AI persona
  - name: calendar
    description: Calendar availability and booking
  - name: voice
    description: Voice agent webhook (Twilio)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /chat/message:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the AI persona
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: Streamed response (text/event-stream)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"

  /chat/history:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get recent chat history for a session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

  /calendar/availability:
    get:
      operationId: getAvailability
      tags: [calendar]
      summary: Get Radhika's available slots for the next 7 days
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"

  /calendar/book:
    post:
      operationId: bookMeeting
      tags: [calendar]
      summary: Book a meeting on Radhika's calendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Meeting booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingConfirmation"

  /voice/inbound:
    post:
      operationId: voiceInbound
      tags: [voice]
      summary: Twilio inbound call webhook — returns TwiML
      responses:
        "200":
          description: TwiML XML response
          content:
            text/xml:
              schema:
                type: string

  /voice/gather:
    post:
      operationId: voiceGather
      tags: [voice]
      summary: Twilio gather webhook — processes speech input and returns TwiML
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/TwilioGatherInput"
      responses:
        "200":
          description: TwiML XML response
          content:
            text/xml:
              schema:
                type: string

  /voice/status:
    post:
      operationId: voiceStatus
      tags: [voice]
      summary: Twilio call status callback
      responses:
        "200":
          description: OK

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ChatMessageInput:
      type: object
      required: [message, sessionId]
      properties:
        message:
          type: string
          description: User message text
        sessionId:
          type: string
          description: Session identifier for conversation continuity

    ChatResponse:
      type: object
      required: [reply, sessionId]
      properties:
        reply:
          type: string
        sessionId:
          type: string
        bookingSlots:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/TimeSlot"

    ChatMessage:
      type: object
      required: [id, sessionId, role, content, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string

    TimeSlot:
      type: object
      required: [start, end, label]
      properties:
        start:
          type: string
          description: ISO 8601 datetime
        end:
          type: string
          description: ISO 8601 datetime
        label:
          type: string
          description: Human-readable label e.g. "Mon Jun 9, 2-3 PM IST"

    AvailabilityResponse:
      type: object
      required: [slots]
      properties:
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"

    BookingInput:
      type: object
      required: [start, end, attendeeEmail, attendeeName]
      properties:
        start:
          type: string
          description: ISO 8601 start datetime
        end:
          type: string
          description: ISO 8601 end datetime
        attendeeEmail:
          type: string
        attendeeName:
          type: string
        summary:
          type: ["string", "null"]

    BookingConfirmation:
      type: object
      required: [eventId, meetLink, start, end]
      properties:
        eventId:
          type: string
        meetLink:
          type: ["string", "null"]
        start:
          type: string
        end:
          type: string
        message:
          type: string

    TwilioGatherInput:
      type: object
      properties:
        SpeechResult:
          type: string
        CallSid:
          type: string
        From:
          type: string