# lib/api-zod/src/generated/types/availabilityResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Radhika Khattar AI Persona API
 * OpenAPI spec version: 0.1.0
 */
import type { TimeSlot } from './timeSlot';

export interface AvailabilityResponse {
  slots: TimeSlot[];
}