# lib/api-zod/src/generated/types/bookingConfirmation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Radhika Khattar AI Persona API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingConfirmation {
  eventId: string;
  /** @nullable */
  meetLink: string | null;
  start: string;
  end: string;
  message?: string;
}