# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Radhika Khattar AI Persona API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  /** ISO 8601 start datetime */
  start: string;
  /** ISO 8601 end datetime */
  end: string;
  attendeeEmail: string;
  attendeeName: string;
  /** @nullable */
  summary?: string | null;
}