# lib/api-zod/src/generated/types/chatMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Radhika Khattar AI Persona API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessageInput {
  /** User message text */
  message: string;
  /** Session identifier for conversation continuity */
  sessionId: string;
}