# lib/api-zod/src/generated/types/timeSlot.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Radhika Khattar AI Persona API
 * OpenAPI spec version: 0.1.0
 */

export interface TimeSlot {
  /** ISO 8601 datetime */
  start: string;
  /** ISO 8601 datetime */
  end: string;
  /** Human-readable label e.g. "Mon Jun 9, 2-3 PM IST" */
  label: string;
}